cutting progress and world canvas

+++ Kitchen_Chaos/Assets/Scripts/Cuttingg Counter.cs
@@ -1,11 +1,25 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using System;
 
 public class CuttinggCounter : BaseCounter
 {
 
-    [SerializeField] CuttingObjectsSO[] cuttingObjectsSOs;
+    [SerializeField] CuttingRecipeSO[] cuttingObjectsSOs;
+
+    [SerializeField] int cuttingProgress;
+
+
+
+    public event EventHandler<OnProgressBarChangedEventArgs> OnProgressBarUpdate;    // declaring event
+
+    public class OnProgressBarChangedEventArgs : EventArgs
+    {
+        public float progressNormalized;
+    }    // making class for sending extra data with the event
+
+
 
     public override void Interact(Player player)
     {
@@ -14,6 +28,17 @@ public class CuttinggCounter : BaseCounter
             if (player.IsKitchenObjectPresent() && HasValidRecepie(player.GetKitchenObjects().GetKitchObjSO()))     //player has kitchen object, that can be cutted
             {
                 player.GetKitchenObjects().SetKitchenObjectParent(this);
+
+                cuttingProgress = 0;     // if player drops, kitchen object
+
+
+                CuttingRecipeSO cuttingRecipeSO = GetCuttingRecipeWithInput(GetKitchenObjects().GetKitchObjSO());
+
+                OnProgressBarUpdate?.Invoke(this, new OnProgressBarChangedEventArgs   //firing event for progress Ui Update
+                {
+                    progressNormalized = (float)cuttingProgress / cuttingRecipeSO.maxCutForCutting 
+                });
+
             }
             else         //both dont have kitchen object
             {
@@ -31,48 +56,64 @@ public class CuttinggCounter : BaseCounter
                 //DO NOTHING
             }
         }
-    }
+    }  // this interaction is the basic interaction to place kitchen objects on the counter
 
 
     public override void InteractAlternate(Player player)
     {
         if (IsKitchenObjectPresent() && HasValidRecepie(GetKitchenObjects().GetKitchObjSO()))    // if any kitchen object is present that can be cutted
         {
-            
+            CuttingRecipeSO cuttingRecipeSO = GetCuttingRecipeWithInput(GetKitchenObjects().GetKitchObjSO());
 
-            KitchenObjectSO outputKitchenObjectSO = GetOutputForInput(GetKitchenObjects().GetKitchObjSO());  //call the function we made to get the output for the recepie
-            Debug.Log(outputKitchenObjectSO);
-            GetKitchenObjects().DestroyItself();  //destroy previous kitchenObject after if we destroy it first we dont have data for input of Recepie
-            KitchenObjects.SpawnKitchenObjectOnParent(outputKitchenObjectSO, this);  //now spawn the output( SLICES )
-            
-        }
-    }
+            cuttingProgress++;
 
-
-    public KitchenObjectSO GetOutputForInput (KitchenObjectSO inputKitchenObjectSO) 
-    {
-        foreach (CuttingObjectsSO cuttingRecepieSO in cuttingObjectsSOs)   // we are looping through every recepie
-        {
-            if (cuttingRecepieSO.input == inputKitchenObjectSO)            // and finding the one which has same SO as player carrying
+            OnProgressBarUpdate?.Invoke(this, new OnProgressBarChangedEventArgs
             {
-                Debug.Log("right recipe found");
-                return cuttingRecepieSO.output;                           // when found return output of that recepie  
+                progressNormalized = (float)cuttingProgress / cuttingRecipeSO.maxCutForCutting
+            });   //firing event for progress Ui Update
+
+            if (cuttingProgress >= cuttingRecipeSO.maxCutForCutting)
+            {
+                KitchenObjectSO outputKitchenObjectSO = GetOutputForInput(GetKitchenObjects().GetKitchObjSO());  //call the function we made to get the output for the recepie
+
+                GetKitchenObjects().DestroyItself();  //destroy previous kitchenObject after if we destroy it first we dont have data for input of Recepie
+
+                KitchenObjects.SpawnKitchenObjectOnParent(outputKitchenObjectSO, this);  //now spawn the output( SLICES )
             }
         }
-        Debug.Log("no matching recepie found");
-        return null;
-    }
+    } // this interaction is used to cut kitchen objects present on cutting counter
+
+
+    public KitchenObjectSO GetOutputForInput (KitchenObjectSO inputSO) 
+    {
+        CuttingRecipeSO cuttingRecipeSO = GetCuttingRecipeWithInput(inputSO);
+        if (cuttingRecipeSO != null)
+        {
+            return cuttingRecipeSO.output;
+        }
+        else
+        {
+            return null;
+        }
+    } // this method return us the output from the recipe SO 
+
 
     private bool HasValidRecepie(KitchenObjectSO inputSO)
     {
-        foreach (var cuttingRecepieSO in cuttingObjectsSOs)
+        CuttingRecipeSO cuttingRecipeSO = GetCuttingRecipeWithInput(inputSO);
+        return cuttingRecipeSO != null;
+    }  // this method return true if it has the recipe of the input SO
+
+
+    private CuttingRecipeSO GetCuttingRecipeWithInput(KitchenObjectSO kitchenObjectSOInput)
+    {
+        foreach (var recipe in cuttingObjectsSOs)
         {
-            if (cuttingRecepieSO.input == inputSO)
+            if (recipe.input == kitchenObjectSOInput)
             {
-                return true;
+                return recipe;
             }
         }
-        Debug.Log("not a valid Kitchen Object");
-        return false;
-    }
+        return null;
+    }    // this method returns us the recipe SO from the kitchen Object Input it has
 }
